+++ src/application/ApplicationList/index.tsx
@@ -1,0 +1,1 @@
+export { ApplicationList } from './ApplicationList';
